% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicfuns.R
\name{connectExfor}
\alias{connectExfor}
\title{Connect to EXFOR MongoDb}
\usage{
connectExfor(...)
}
\arguments{
\item{...}{see function \code{mongo} of package \code{mongolite}}
}
\value{
Returns a list with the function described below.
}
\description{
Connect to EXFOR MongoDb
}
\section{Methods}{

\describe{
  \item{\code{iterate(query = '{}', fields = '{"_id":0}', sort = '{}', skip = 0, limit = 0)}}{Runs query and returns iterator to read single records one-by-one.}
  \item{\code{find(query, expr, no.dollar, ...}}{Runs query and returns a data.table constructed from \code{expr} evaluated within the documents in the collection.}
}
}

