% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicfuns.R
\name{exforIterator}
\alias{exforIterator}
\title{Create EXFOR Iterator}
\usage{
exforIterator(...)
}
\arguments{
\item{...}{arguments passed to \code{iterate} function of \code{mongolite} package,
see \link[mongolite]{mongo}}
}
\value{
list with functions \code{getCur()} and \code{getNext()}
}
\description{
Create EXFOR Iterator
}
