% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfuns.R
\name{makeQueryStr}
\alias{makeQueryStr}
\title{Make MongoDb Query String}
\usage{
makeQueryStr(query, auto.quote = TRUE)
}
\arguments{
\item{query}{the query expression}

\item{auto.quote}{should field names be automatically quoted (default is TRUE)}
}
\value{
a character string containing the JSON object for the query
}
\description{
Make MongoDb Query String
}
